Refactor identity-iota layout
