+++ identity-account/src/updates/update.rs
@@ -18,8 +18,8 @@ use identity_did::verification::MethodScope;
 use identity_did::verification::MethodType;
 use identity_iota::did::IotaDID;
 use identity_iota::did::IotaDIDUrl;
-use identity_iota::did::IotaDocument;
-use identity_iota::did::IotaVerificationMethod;
+use identity_iota::document::IotaDocument;
+use identity_iota::document::IotaVerificationMethod;
 use identity_iota::tangle::UPDATE_METHOD_TYPES;
 
 use crate::account::Account;
