+++ identity-iota/src/chain/document_history.rs
@@ -6,9 +6,9 @@ use std::ops::Deref;
 
 use crate::chain::DiffChain;
 use crate::chain::IntegrationChain;
-use crate::did::DocumentDiff;
 use crate::did::IotaDID;
-use crate::did::IotaDocument;
+use crate::document::DocumentDiff;
+use crate::document::IotaDocument;
 use crate::error::Result;
 use crate::tangle::Client;
 use crate::tangle::Message;
