--- identity-iota/src/did/did/mod.rs
@@ -1,9 +1,0 @@
-// Copyright 2020-2021 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-mod iota_did;
-mod segments;
-
-pub use self::iota_did::IotaDID;
-pub use self::iota_did::IotaDIDUrl;
-pub use self::segments::Segments;
