+++ identity-iota/src/did/mod.rs
@@ -1,19 +1,10 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#[macro_use]
+pub use self::iota_did::IotaDID;
+pub use self::iota_did::IotaDIDUrl;
+pub use self::segments::Segments;
+
+mod iota_did;
 mod macros;
-
-#[allow(clippy::module_inception)]
-mod did;
-mod doc;
-
-pub use self::did::IotaDID;
-pub use self::did::IotaDIDUrl;
-pub use self::did::Segments;
-pub use self::doc::DocumentDiff;
-pub use self::doc::IotaDocument;
-pub use self::doc::IotaDocumentSigner;
-pub use self::doc::IotaDocumentVerifier;
-pub use self::doc::IotaVerificationMethod;
-pub use self::doc::Properties;
+mod segments;
