+++ identity-iota/src/did/segments.rs
