+++ identity-iota/src/document/iota_document.rs
@@ -8,7 +8,6 @@ use core::fmt::Display;
 use core::fmt::Formatter;
 use core::fmt::Result as FmtResult;
 
-use identity_did::verification::MethodRelationship;
 use serde::Serialize;
 
 use identity_core::common::Object;
@@ -35,17 +34,18 @@ use identity_did::verifiable::DocumentVerifier;
 use identity_did::verifiable::Properties as VerifiableProperties;
 use identity_did::verification::MethodQuery;
 use identity_did::verification::MethodRef;
+use identity_did::verification::MethodRelationship;
 use identity_did::verification::MethodScope;
 use identity_did::verification::MethodType;
 use identity_did::verification::MethodUriType;
 use identity_did::verification::TryMethod;
 use identity_did::verification::VerificationMethod;
 
-use crate::did::DocumentDiff;
 use crate::did::IotaDID;
 use crate::did::IotaDIDUrl;
-use crate::did::IotaVerificationMethod;
-use crate::did::Properties as BaseProperties;
+use crate::document::DocumentDiff;
+use crate::document::IotaVerificationMethod;
+use crate::document::Properties as BaseProperties;
 use crate::error::Error;
 use crate::error::Result;
 use crate::tangle::MessageId;
@@ -89,7 +89,7 @@ impl IotaDocument {
   ///
   /// ```
   /// # use identity_core::crypto::KeyPair;
-  /// # use identity_iota::did::IotaDocument;
+  /// # use identity_iota::document::IotaDocument;
   /// #
   /// // Create a DID Document from a new Ed25519 keypair.
   /// let keypair = KeyPair::new_ed25519().unwrap();
@@ -114,7 +114,7 @@ impl IotaDocument {
   ///
   /// ```
   /// # use identity_core::crypto::KeyPair;
-  /// # use identity_iota::did::IotaDocument;
+  /// # use identity_iota::document::IotaDocument;
   /// # use identity_iota::tangle::Network;
   /// #
   /// // Create a new DID Document for the devnet from a new Ed25519 keypair.
@@ -376,8 +376,8 @@ impl IotaDocument {
   /// Returns an iterator over all [`IotaVerificationMethods`][IotaVerificationMethod] in the DID Document.
   pub fn methods(&self) -> impl Iterator<Item = &IotaVerificationMethod> {
     self.document.methods().map(|m|
-        // SAFETY: Validity of verification methods checked in `IotaVerificationMethod::check_validity`.
-        unsafe { IotaVerificationMethod::new_unchecked_ref(m) })
+      // SAFETY: Validity of verification methods checked in `IotaVerificationMethod::check_validity`.
+      unsafe { IotaVerificationMethod::new_unchecked_ref(m) })
   }
 
   /// Adds a new [`IotaVerificationMethod`] to the document in the given [`MethodScope`].
@@ -814,34 +814,17 @@ mod tests {
 
   use identity_core::common::Value;
   use identity_core::convert::FromJson;
-  use identity_core::convert::SerdeInto;
   use identity_core::crypto::merkle_key::Sha256;
   use identity_core::crypto::KeyCollection;
-  use identity_core::crypto::KeyPair;
   use identity_core::crypto::KeyType;
-  use identity_core::crypto::PrivateKey;
-  use identity_core::crypto::PublicKey;
   use identity_core::utils::encode_b58;
   use identity_did::did::CoreDID;
-  use identity_did::did::CoreDIDUrl;
   use identity_did::did::DID;
-  use identity_did::document::CoreDocument;
-  use identity_did::service::Service;
   use identity_did::verification::MethodData;
-  use identity_did::verification::MethodRef;
-  use identity_did::verification::MethodScope;
-  use identity_did::verification::MethodType;
-  use identity_did::verification::VerificationMethod;
 
-  use crate::did::did::IotaDID;
-  use crate::did::doc::iota_document::Properties;
-  use crate::did::doc::IotaDocument;
-  use crate::did::doc::IotaVerificationMethod;
-  use crate::did::IotaDIDUrl;
-  use crate::tangle::MessageId;
   use crate::tangle::Network;
-  use crate::tangle::TangleRef;
-  use crate::Error;
+
+  use super::*;
 
   const DID_ID: &str = "did:iota:HGE4tecHWL2YiZv5qAGtH7gaeQcaz2Z1CR15GWmMjY1M";
   const DID_METHOD_ID: &str = "did:iota:HGE4tecHWL2YiZv5qAGtH7gaeQcaz2Z1CR15GWmMjY1M#sign-0";
