+++ identity-iota/src/lib.rs
@@ -2,37 +2,36 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #![cfg_attr(docsrs, feature(doc_cfg, extended_key_value_attributes))]
-#![cfg_attr(docsrs, cfg_attr(docsrs, doc = include_str!("../README.md")))]
+#![cfg_attr(docsrs, cfg_attr(docsrs, doc = include_str ! ("../README.md")))]
 #![cfg_attr(not(docsrs), doc = "")]
 #![allow(clippy::upper_case_acronyms)]
 #![warn(
-  rust_2018_idioms,
-  unreachable_pub,
-  // missing_docs,
-  rustdoc::missing_crate_level_docs,
-  rustdoc::broken_intra_doc_links,
-  rustdoc::private_intra_doc_links,
-  rustdoc::private_doc_tests,
-  clippy::missing_safety_doc,
-  // clippy::missing_errors_doc,
+rust_2018_idioms,
+unreachable_pub,
+// missing_docs,
+rustdoc::missing_crate_level_docs,
+rustdoc::broken_intra_doc_links,
+rustdoc::private_intra_doc_links,
+rustdoc::private_doc_tests,
+clippy::missing_safety_doc,
+// clippy::missing_errors_doc,
 )]
 
 #[macro_use]
 extern crate lazy_static;
-
 #[macro_use]
 extern crate log;
-
 #[macro_use]
 extern crate serde;
+
+pub use self::error::Error;
+pub use self::error::Result;
 
 mod resolver;
 
 pub mod chain;
 pub mod credential;
 pub mod did;
+pub mod document;
 pub mod error;
 pub mod tangle;
-
-pub use self::error::Error;
-pub use self::error::Result;
